feat: move chainList to static
