+++ libs/static/src/chains.ts
@@ -1,4 +1,4 @@
-import { ChainProperties, Config } from './types'
+import type { ChainProperties, Config, Option, Prefix } from './types'
 
 export const toChainProperty = (
   ss58Format: number,
@@ -33,4 +33,30 @@ export const CHAINS: Config<ChainProperties> = {
   ),
   movr: toChainProperty(1285, 18, 'MOVR', 'https://moonriver.subscan.io/'),
   glmr: toChainProperty(1284, 18, 'GLMR', 'https://moonbeam.subscan.io/'),
+}
+
+export const chainPrefixes: Prefix[] = [
+  'bsx',
+  'ksm',
+  'snek',
+  'movr',
+  'glmr',
+  'rmrk2',
+]
+
+export const chainInfo = {
+  bsx: 'basilisk',
+  ksm: 'kusama',
+  snek: 'snek',
+  movr: 'moonriver',
+  glmr: 'moonbeam',
+  rmrk2: 'rmrk',
+}
+
+export const chainList = (): Option[] => {
+  return chainPrefixes.map((prefix) => ({
+    info: chainInfo[prefix],
+    text: chainInfo[prefix],
+    value: prefix === 'ksm' ? 'rmrk' : prefix,
+  }))
 }
