+++ utils/chain.ts
@@ -1,6 +1,8 @@
 import { ENDPOINTS } from '@kodadot1/vuex-options'
-import { NAMES } from '@kodadot1/static'
-import type { Option, Prefix } from '@kodadot1/static'
+import { chainList } from '@kodadot1/static'
+import { disableChainListOnProductionEnv } from './constants'
+
+import type { Option } from '@kodadot1/static'
 
 const prefixes: Record<string, number> = {
   polkadot: 0,
@@ -84,30 +86,14 @@ export const getChainNameByPrefix = (prefix: string) => {
   return prefix
 }
 
-// TODO: move @kodadot1/static under nft-gallery
-// https://github.com/kodadot/packages/issues/112
-export const chainPrefixes: Prefix[] = [
-  'bsx',
-  'ksm',
-  'snek',
-  'movr',
-  'glmr',
-  'rmrk2',
-]
-
-const infos = {
-  bsx: 'basilisk',
-  ksm: 'kusama',
-  snek: 'snek',
-  movr: 'moonriver',
-  glmr: 'moonbeam',
-  rmrk2: 'rmrk',
-}
-
 export const availablePrefixes = (): Option[] => {
-  return chainPrefixes.map((prefix) => ({
-    info: infos[prefix],
-    text: NAMES[prefix],
-    value: prefix === 'ksm' ? 'rmrk' : prefix,
-  }))
+  const chains = chainList()
+
+  if (window.location.hostname === 'kodadot.xyz') {
+    return chains.filter(
+      (chain) => !disableChainListOnProductionEnv.includes(String(chain.value))
+    )
+  }
+
+  return chains
 }
